+++ packages/cursorless-org-docs/src/docs/components/flattenHighlights.ts
